perbaikan tipe data product, dan tampilan price

+++ database/migrations/2024_12_07_012010_create_product_products_table.php
@@ -12,11 +12,11 @@ class CreateProductProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('product_description')->nullable();
-            $table->decimal('product_stock',8,2)->nullable();
+            $table->integer('product_stock')->nullable();
             $table->string('address')->nullable();
             $table->string('size')->nullable();
             $table->decimal('product_price',12,2)->nullable();
-            $table->decimal('discount',8,2)->nullable();
+            $table->decimal('discount',8,1)->nullable();
             $table->string('consist_of')->nullable();
             $table->string('main_picture')->nullable();
             $table->timestamps();
